+++ src/etui/config.py
@@ -10,11 +10,10 @@ from platformdirs import user_config_dir
 from etui.file_utils import ROOT_PATH, PYTHON_UV
 
 DEFAULT_CONFIG_DIR = ROOT_PATH / "config_defaults"
-
 USER_CONFIG_DIR = Path(user_config_dir("etui"))
-
 SETTINGS_FILE = "settings.toml"
 SCRIPT_FOLDERS_FILE = "script_folders.toml"
+
 
 @dataclass
 class ScriptFolder:
@@ -22,19 +21,18 @@ class ScriptFolder:
     path: Path
     executable: Path = PYTHON_UV
     cwd: Path = ROOT_PATH
-    file_extension: str = ".py"
+    file_extension: str = "*.py"
     exclude_start: tuple[str] = ("__",)
 
 
 def ensure_user_configs():
+    """Ensures that the user config files are present."""
     USER_CONFIG_DIR.mkdir(parents=True, exist_ok=True)
-
     for name in (SETTINGS_FILE, SCRIPT_FOLDERS_FILE):
-        user = USER_CONFIG_DIR / name
-        default = DEFAULT_CONFIG_DIR / name
-
-        if not user.exists():
-            shutil.copy(default, user)
+        user_config = USER_CONFIG_DIR / name
+        default_config = DEFAULT_CONFIG_DIR / name
+        if not user_config.exists():
+            shutil.copy(default_config, user_config)
 
 
 def load_toml(path: Path) -> dict:
@@ -46,15 +44,18 @@ def load_settings() -> dict:
     return load_toml(USER_CONFIG_DIR / SETTINGS_FILE)
 
 
-def load_script_folders() -> list[dict]:
-    data = load_toml(USER_CONFIG_DIR / SCRIPT_FOLDERS_FILE)
-    folders = data.get("folders", [])
-
-    for folder in folders:
-        folder["path"] = (ROOT_PATH / folder["path"]).resolve()
-        folder["cwd"] = (ROOT_PATH / folder["cwd"]).resolve()
-        folder["python"] = (ROOT_PATH / folder["python"]).resolve()
-
+def load_script_folders() -> dict[str, ScriptFolder]:
+    toml_folders = load_toml(USER_CONFIG_DIR / SCRIPT_FOLDERS_FILE).get("folders", [])
+    folders = {}
+    for folder in toml_folders:
+        folders[folder["name"]] = ScriptFolder(
+            folder["name"],
+            Path(folder["path"]),
+            Path(folder["executable"]),
+            folder["cwd"],
+            "*" + folder["file_extension"],
+            folder["exclude_start"],
+        )
     return folders