+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project (more or less) adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.0] - 2026-01-12
+
+### Added:
+- Default settings and script folders files
+- Settings and script folders file moved outside repo to default config directory
+- Loading script folders from the script folders file
+
+### Removed
+- Hardcoded ScriptFolder
+
 
 ## [0.1.2] - 2026-01-07